ml-project/requirements.txt
PyYAML>=6.0.0
lightgbm>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
flask>=3.0.0
category_encoders>=2.6.0
optuna>=3.4.0
xgboost>=2.0.0
